main: add check subcommand.
Check command fails if the current lock would be affected by `prune`.


The `check` command takes no output options since it will not write
any changes, but just check that prunning results in the same *hits*
(FlakeNodeVisits) as those computed for the current lock file.


After computing hits for the _current_ and _pruned_ lock files,
the `check` command panics if they are different.
panic! was used as it would terminate and write to stderr.
(elog! eprintln! macros dont write to stderr)


Closes #4