nix: add treewide formatter
Introduce a simple script that handles exactly two cases:

- Formatting Nix code from *stdin*, e.g., when the editor is
  configured to use `nix fmt -- --` as the default Nix formatter.
- Formatting the entire source tree from the current working directory.

In the latter case, both `nixfmt .` and `cargo fmt` are invoked.

It is intentional that `--all` is not passed to `cargo fmt`.

Arbitrary file paths are not supported, as indicated by the usage text,
which is printed to *stderr* if an unexpected argument is provided.