main: Command: Prune: add option relax-inputs
This new option can be specified for many input names. The names used in
this option's values will not be subject to skipping if the transitive
input of the same name is from a different branch.

Prior to #15 (a fix for #14) the default behavior was to ignore the
`original` attribute of the locked input, meaning that input
substitutions were performed by-name only, and did not take into account
that some dependencies may add inputs with the same name as exist in the
root node.